--- source/App/Syntax/CommandOptions.cs
@@ -1,14 +1,0 @@
-﻿namespace Petrichor.App.Syntax
-{
-	public readonly struct CommandOptions
-	{
-		public const string DefaultCommandOptionHelp = "--help";
-		public const string ShortcutScriptLogModeArgumentAll = "all";
-		public const string ShortcutScriptLogModeArgumentConsoleOnly = "consoleOnly";
-		public const string ShortcutScriptLogModeArgumentFileOnly = "fileOnly";
-		public const string ShortcutScriptOptionInputFile = "--inputFile";
-		public const string ShortcutScriptOptionLogFile = "--logFile";
-		public const string ShortcutScriptOptionLogMode = "--logMode";
-		public const string ShortcutScriptOptionOutputFile = "--outputFile";
-	}
-}
