+++ source/ShortcutScriptGeneration/Syntax/Commands.cs
@@ -1,0 +1,7 @@
+﻿namespace Petrichor.ShortcutScriptGeneration.Syntax
+{
+	public readonly struct Commands
+	{
+		public const string ModuleCommand = "generateTextShortcutScript";
+	}
+}
